invitation

+++ Household_Budgeter/Controllers/HouseholdsController.cs
@@ -7,21 +7,31 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using Household_Budgeter.Models;
-
+using Microsoft.AspNet.Identity;
 
 namespace Household_Budgeter.Controllers
 {
+    [Authorize]
     public class HouseholdsController : Controller
     {
         private ApplicationDbContext db = new ApplicationDbContext();
 
-        // GET: Households1
         public ActionResult Index()
         {
-            return View(db.Households.ToList());
+
+
+            var user = db.Users.Find(User.Identity.GetUserId());
+            Household household = db.Households.Find(user.HouseholdId);
+            if (household == null)
+            {
+                return RedirectToAction("Create", "Households");
+            }
+
+
+            return View(db.Households.Find(user.HouseholdId));
         }
 
-        // GET: Households1/Details/5
+        // GET: Households/Details/5
         public ActionResult Details(int? id)
         {
             if (id == null)
@@ -34,15 +44,15 @@ namespace Household_Budgeter.Controllers
                 return HttpNotFound();
             }
             return View(household);
-        }        
+        }
 
-        // GET: Households1/Create
+        // GET: Households/Create
         public ActionResult Create()
         {
             return View();
         }
 
-        // POST: Households1/Create
+        // POST: Households/Create
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
@@ -51,15 +61,29 @@ namespace Household_Budgeter.Controllers
         {
             if (ModelState.IsValid)
             {
-                db.Households.Add(household);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
+                var user = db.Users.Find(User.Identity.GetUserId());
 
+                if (user.HouseholdId == null)
+                {
+                    db.Households.Add(household);
+                    db.SaveChanges();
+
+                    var getHousehold = db.Households.FirstOrDefault(h => h.Name == household.Name);
+                    user.HouseholdId = getHousehold.Id;
+
+                    db.SaveChanges();
+
+                    return RedirectToAction("Index", new { id = getHousehold.Id });
+                }
+                else
+                {
+                    return RedirectToAction("Index", new { id = user.Id });
+                }
+            }
             return View(household);
         }
 
-        // GET: Households1/Edit/5
+        // GET: Households/Edit/5
         public ActionResult Edit(int? id)
         {
             if (id == null)
@@ -74,7 +98,7 @@ namespace Household_Budgeter.Controllers
             return View(household);
         }
 
-        // POST: Households1/Edit/5
+        // POST: Households/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
@@ -90,7 +114,7 @@ namespace Household_Budgeter.Controllers
             return View(household);
         }
 
-        // GET: Households1/Delete/5
+        // GET: Households/Delete/5
         public ActionResult Delete(int? id)
         {
             if (id == null)
@@ -105,12 +129,30 @@ namespace Household_Budgeter.Controllers
             return View(household);
         }
 
-        // POST: Households1/Delete/5
+        public ActionResult Leave()
+        {
+            var user = db.Users.Find(User.Identity.GetUserId());
+            user.HouseholdId = null;
+
+            db.SaveChanges();
+            return RedirectToAction("Create", "Household");
+        }
+
+        // POST: Households/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            var user = db.Users.Find(User.Identity.GetUserId());
+            BankAccount bankaccount = db.BankAccounts.FirstOrDefault(x => x.Id == id);
+
             Household household = db.Households.Find(id);
+
+            if (!household.Members.Contains(user))
+            {
+                return RedirectToAction("Unauthorized", "Error");
+            }
+
             db.Households.Remove(household);
             db.SaveChanges();
             return RedirectToAction("Index");
